style: 유저 사용량 grid table 컴포넌트 분리
